Changed documentation for "addExprData"
Found an issue where documentation examples must be runnable R commands. In order to use the funcion
you are documenting in the documentation example it is necesarry to export the function. This can cause problems because it makes functions visible to the users which they were not intended to use. A work around to this is to just use the ":::" operator to acces functions which are not exported by packages. Using this approach it is unnecessary to export every function you would like to include an example with.