Updated documentation
The documentation was updated and regenerated using roxygen. This was done to create working examples for functions which I did not wish to export. To access these functions which were not exported I used the ":::" operator. Checked and Built. No Errors. 1 Warning. 2 Notes.

Warning is because there are shiny functions in some files which I have not prefixed with "shiny::"